+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/templatest/compare/v0.7.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Add `templatest.__version__`
 - Add py.typed
 
 [0.7.0](https://github.com/jshwi/templatest/releases/tag/v0.7.0) - 2022-10-30